Use DFAs to get the intersection of two subexpr.
We use dk.brics.automaton, a DFA library. Because the library only
accepts regexes as strings, we need to reconstruct them from RegexTRree.
This is what ReconstructionVisitor is for.

brics and Java regexes do not use the same syntax. A small part of the
syntax is translated, but not all features work.